Actually download new videos again
SVT changed so that the playlists downloaded for a bitrate could be
a complete URL instead of just part of a chunk. This new naming scheme
changed so that downloads didn't complete fully on new videos.

So now the script always use the complete URL from the playlist instead
of making up its own URL based on off how it ought to be. Thus this
should be more future proof.

Fingers crossed. :)

Actual changes
- Internal refactoring because of part naming changes
  - @first_part / @last_part removed and an array of all parts added
    instead
  - Set the base_url from the part if set instead of guessing based off
    the playlist (should be the same, but who knows what'll happen)
- svt-recorder program now catches more errors and retries fetching,
  probably Keep-Alive was being closed by the server. Not really sure
  why but files seems to be working.